requirements.txt
pandas==2.1.4
tabula_py==2.9.0
